Rename BaseEntity to AbstractEntity, move both it and AbstractAction to top level
