+++ Civi/Api4/Action/Create.php
@@ -27,6 +27,7 @@
 
 namespace Civi\Api4\Action;
 
+use Civi\Api4\AbstractAction;
 use Civi\Api4\Result;
 
 /**
