+++ Civi/Api4/Action/GetActions.php
@@ -28,6 +28,7 @@
 namespace Civi\Api4\Action;
 
 use Civi\API\Exception\NotImplementedException;
+use Civi\Api4\AbstractAction;
 use Civi\Api4\Result;
 use Civi\Api4\Utils\ReflectionUtils;
