+++ Civi/Api4/Action/GetFields.php
@@ -27,6 +27,7 @@
 
 namespace Civi\Api4\Action;
 
+use Civi\Api4\AbstractAction;
 use Civi\Api4\Service\Spec\SpecGatherer;
 use Civi\Api4\Result;
 use Civi\Api4\Service\Spec\SpecFormatter;
