+++ tests/phpunit/Entity/ConformanceTest.php
@@ -2,7 +2,7 @@
 
 namespace Civi\Test\Api4\Entity;
 
-use Civi\Api4\Entity\BaseEntity;
+use Civi\Api4\AbstractEntity;
 use Civi\Test\Api4\UnitTestCase;
 
 /**
@@ -109,7 +109,7 @@ class ConformanceTest extends UnitTestCase {
 
     $this->hookClass->setMock($this);
     // get list of all the entities we know about and loop over them:
-    $entities = BaseEntity::get()
+    $entities = AbstractEntity::get()
       ->setCheckPermissions(FALSE)
       ->execute();
     foreach ($entities as $entity) {
